AI/requirements.txt
click==8.1.7
deepface==0.0.93
huggingface-hub==0.25.1
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
six==1.16.0
tensorflow==2.17.0
tqdm==4.66.5
transformers==4.45.2
playsound==1.3.0  ## python <= 3.10
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
openai==1.3.3


pymongo~=4.10.1
num2words~=0.5.12
jamo~=0.4.1
cn2an~=0.5.22
jieba~=0.42.1
pypinyin~=0.50.0
g2p-en~=2.1.0
mecab-python3~=1.0.9
pykakasi~=2.2.1
inflect~=7.0.0
gruut~=2.2.3
torch~=2.4.1
librosa~=0.9.1
torchaudio~=2.4.1
gradio~=4.44.1
scipy~=1.14.1
loguru~=0.7.2
numba~=0.60.0
cached_path~=1.6.3
PyAudio~=0.2.14
websockets~=12.0
pytesseract~=0.3.13
onnxruntime~=1.19.2
imread_from_url~=0.1.3
pyzbar~=0.1.9
fastapi~=0.115.2